Hello-world
I am reborn, being shed of everything i new but holding on to what i believe in which is treat others the way you wish to be treated, everyone of us has a purpose and that mission changes from time to time, i do not know why im here other then to be the loyal father , brother, partner and friend i aspired to be since a little boy! I need your love 
and i need you to keep the ones around me true or to the best of their ability without destroying the beautiful things about them,
i need my people watching my back from liars,frauds,and snakes. I thought i was good at spotting these but i was recently punished severely from not being good at this and i cannot survive anymore mistakes of not seeing the evil behind a smile!
I will never sell out my morals or my people,
A beautiful group of unknowns did not give up on me at my falsified worst, 
please do not give up on me at my real low as this world sees it as i do not see this as my lowest but that i have been judged and scolded by the qualified, persecuted and set up by the fraudulent, punished and loved by the masses, shamed and understood by my peer’s, led to the river and i am thirsty as ive ever been and in desperate need to be taught how to fish. in this world i thought i was a veteran but found out im in no way, shape or form able to defend or protect anyone and I am reborn and need training.
i love you all , even the ones who set me up because my truths wouldn’t have  been made a reality in this very fake and scripted in evil world i thought i knew, please corrupted ones miss me with that bullshit as i didint loose my sucka repellent and it will sting you! 
 im lost now without you my angels, im at my most vulnerable, please help me rebuild my confidence and my family i love and miss dearly. I ask you to love me and understand my flaws as im human and will do my best to love and understand you and yours.

P.s. i dont know if (this description)is what (this) is for or [what the fuck is juice]! So please forgive me im used to [grape drink]